feat(share): 默认 PNG/1.5x 并收紧分享卡版式
默认导出改为 PNG 与 1.5x 清晰度，卡片逻辑宽 600 并放大正文字号；备份字段与设置项对齐。

+++ lib/services/post_share_service.dart
@@ -83,7 +83,7 @@ class _SharePreviewSheetState extends ConsumerState<_SharePreviewSheet> {
   String _statusMessage = '';
 
   late ShareImageFormat _shareImageFormat;
-  late int _sharePixelRatio;
+  late double _sharePixelRatio;
 
   @override
   void initState() {
@@ -216,7 +216,7 @@ class _SharePreviewSheetState extends ConsumerState<_SharePreviewSheet> {
     final renderObject = _captureKey.currentContext?.findRenderObject()
         as RenderRepaintBoundary?;
     if (renderObject == null) return null;
-    final pixelRatio = _sharePixelRatio.toDouble();
+    final pixelRatio = _sharePixelRatio;
 
     return _captureFromRepaint(renderObject, pixelRatio);
   }
